Use single module to driver, container, and blob for type hinting
