+++ cloudstorage/drivers/google.py
@@ -23,10 +23,7 @@ from google.cloud.storage.blob import Blob as GoogleBlob
 from google.cloud.storage.bucket import Bucket
 from inflection import underscore
 
-from cloudstorage import messages
-from cloudstorage.blob import Blob
-from cloudstorage.container import Container
-from cloudstorage.driver import Driver
+from cloudstorage import Blob, Container, Driver, messages
 from cloudstorage.exceptions import (
     CloudStorageError,
     IsNotEmptyError,
