+++ cloudstorage/drivers/local.py
@@ -15,10 +15,7 @@ import itsdangerous
 import xattr
 from inflection import underscore
 
-from cloudstorage import messages
-from cloudstorage.blob import Blob
-from cloudstorage.container import Container
-from cloudstorage.driver import Driver
+from cloudstorage import Blob, Container, Driver, messages
 from cloudstorage.exceptions import (
     CloudStorageError,
     IsNotEmptyError,
