+++ cloudstorage/drivers/microsoft.py
@@ -24,10 +24,7 @@ from azure.storage.blob.models import ContentSettings
 from azure.storage.blob.models import BlobPermissions
 from inflection import underscore
 
-from cloudstorage import messages
-from cloudstorage.blob import Blob
-from cloudstorage.container import Container
-from cloudstorage.driver import Driver
+from cloudstorage import Blob, Container, Driver, messages
 from cloudstorage.exceptions import (
     NotFoundError,
     CloudStorageError,
