+++ docs/api/blob.rst
@@ -1,7 +1,7 @@
 Blob
 ----
 
-.. autoclass:: cloudstorage.blob.Blob
+.. autoclass:: cloudstorage.base.Blob
     :members:
     :special-members: __len__
     :member-order: bysource
