+++ docs/api/container.rst
@@ -1,7 +1,7 @@
 Container
 ---------
 
-.. autoclass:: cloudstorage.container.Container
+.. autoclass:: cloudstorage.base.Container
     :members:
     :special-members: __iter__, __contains__, __len__
     :member-order: bysource
