+++ docs/api/driver.rst
@@ -1,7 +1,7 @@
 Driver
 ------
 
-.. autoclass:: cloudstorage.driver.Driver
+.. autoclass:: cloudstorage.base.Driver
     :members:
     :special-members: __iter__, __contains__, __len__
     :member-order: bysource
