+++ tests/test_drivers.py
@@ -1,5 +1,5 @@
 from cloudstorage import DriverName, get_driver, get_driver_by_name
-from cloudstorage.driver import Driver
+from cloudstorage.base import Driver
 
 
 def test_get_driver():
